рефакторинг 2(сложные компоненты)
