+++ src/view/event-add-view.js
@@ -3,59 +3,21 @@ import { destinations } from '../utils/destinations.js';
 import { wayPointTypes } from '../utils/waypointTypes.js';
 import { generateImages } from '../utils/functions.js';
 import AbstractView from './abstract-view.js';
+import { createOffersSegmentMarkup, createWaypointTypesMarkup } from '../utils/forms.js';
 
 const createEventAddTemplate = (point) => {
-  const {offers, description } = point;
+  const {offers, description} = point;
   const waypointType = 'taxi';
   const templateDatetime = dayjs().add(14, 'day').hour(0).minute(0).format('DD/MM/YY HH:mm');
 
-  const createOfferMarkup = (offer) => {
-    const offerName = offer.name;
-    const offerPrice = offer.price;
-    const offerType = offer.type;
-    return `<div class="event__available-offers">
-                      <div class="event__offer-selector">
-                        <input class="event__offer-checkbox  visually-hidden" id="event-offer-${offerType}-1" type="checkbox" name="event-offer-${offerType}" >
-                        <label class="event__offer-label" for="event-offer-name-1">
-                          <span class="event__offer-title">${offerName}</span>
-                          &plus;&euro;&nbsp;
-                          <span class="event__offer-price">${offerPrice}</span>
-                        </label>
-                      </div>
-    `;
-  };
-
-  const createOfferListMarkup = (addableOffers) => {
-    if (addableOffers.length !== 0){
-      return `<section class="event__section  event__section--offers">
-                    <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-                    ${offers.map(createOfferMarkup).join('')}</section>`;
-    }
-    return '';
-  };
 
   const images = generateImages();
 
-  const createWaypointTypesMarkup = (types =  wayPointTypes(), chosenWaypointType) => {
-    const createType = (currentType) => {
-      const isChecked = currentType === chosenWaypointType ? 'checked=""' : '';
-      const label = currentType.charAt(0).toUpperCase() + currentType.slice(1);
-      return `<div class="event__type-item">
-                          <input id="event-type-${currentType}-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="${currentType}" ${isChecked}>
-                          <label class="event__type-label  event__type-label--${currentType}" for="event-type-${currentType}-1">${label}</label>
-                        </div>`;
-    };
-    return types.map(createType).join('');
-  };
-
-  const createOptionsLocations = (city) => (`<option value="${city}"></option>`);
-
-  const createPhotoMarkup = (photo) => (`<img className="event__photo" src="${photo}">`);
 
   const waypointTypesMarkup = createWaypointTypesMarkup(wayPointTypes(), waypointType);
-  const addableOffersMarkup = createOfferListMarkup(offers);
-  const imagesList = images.map(createPhotoMarkup).join('');
-  const optionsLocations = destinations().map(createOptionsLocations).join('');
+  const addableOffersMarkup = createOffersSegmentMarkup(offers);
+  const imagesList = images.map((x) => (`<img className="event__photo" src="${x}">`)).join('');
+  const optionsLocations = destinations().map((x) => (`<option value="${x}"></option>`)).join('');
   const waypointTypeLabel = waypointType.charAt(0).toUpperCase() + waypointType.slice(1);
